// src/enums/selector.ts
/**
 * Selectors
 */
export enum Selector {
    USERNAME_FIELD = '#txt_usuario',
    PASSWORD_FIELD = '#txt_clave',
    BTN = '.btn',
    BUTTON = 'button',
    SERVICES_COMBOBOX = 'select#cmb_servicio',
    SERVICES_TRAINING_GROUND_ID = '37',
    CALENDAR = '#divCalendario',
    ACTIVE_DAY_LIST_ELEMENT = 'li:has(span.active)',
    ROWS = 'table > tbody > tr',
    FIRST_COLUMN = 'td:first-child',
    CONFIRM_BOOKING_MODAL = '#md_modal_confirmar',
    ACCEPT_BUTTON = '#md_modal_confirmar_btn_aceptar',
    ERROR_MESSAGE_LABEL = 'label#md_modal_confirmar_lbl_mensaje.error',
    DISMISS_BUTTON = 'button.btn.btn-outline-secondary[title="Cancelar"]',
    CLOSE_BUTTON = '#md_modal_mensaje_cerrar',
    TIMETABLE_MODAL = '#md_modal_horas',
    SPAN = 'span'

}